docs(generators/*/templates): fix `@static` declarations for `@property`-s
The best practice of the day is that a `@property` should not be
`@static` unless it really is a property on the class itself, that is,
not on its prototype.

This leaves singleton objects in a bit of a gray area.

Strictly, a singleton object still is an instance of a class and so its
`@property`-s should not be declared `@static`.
On the other hand its `@class` *does* get  declared `@static` itself in
order to signify that the module's export is, in fact, the object and
not the `class`. And so, one could argue a `@property` should be
`@static` as well; but then, shouldn't its `@method`-s as well?

And then there's `Settings` its properties are better defined as
`@attribute`-s because they can only be accessed through it's get/set
API.